/ameliorer_etat_des_champs: Auto stash before rebase of "origin/bugfix/ameliorer_etat_des_champs"
