+++ packages/modul-components/src/components/datepicker/datepicker.ts
@@ -148,7 +148,7 @@ export class MDatepicker extends ModulVue {
     }
 
     public set open(open: boolean) {
-        if (this.as<InputState>().active) {
+        if (!this.as<InputState>().isDisabled && !this.as<InputState>().isReadonly) {
             this.internalOpen = open;
         }
     }
@@ -275,7 +275,10 @@ export class MDatepicker extends ModulVue {
     }
 
     public togglePopup(event: Event): void {
-        this.open = !this.open;
+        if (!this.as<InputState>().isDisabled && !this.as<InputState>().isReadonly) {
+            this.open = !this.open;
+        }
+
         // stop event propagation to parent.
         event.stopPropagation();
     }
@@ -285,7 +288,7 @@ export class MDatepicker extends ModulVue {
     }
 
     public onKeydown(event: KeyboardEvent): void {
-        if (this.as<InputStateMixin>().active) {
+        if (!this.as<InputStateMixin>().isDisabled && !this.as<InputStateMixin>().isReadonly) {
             if (event.key === 'Tab') {
                 // close popop if open and tab key is pressed (accessibility)
                 if (this.open) {
@@ -306,11 +309,11 @@ export class MDatepicker extends ModulVue {
 
     // override from InputManagement
     public onFocus(event: FocusEvent): void {
-        if (!this.open) { // open on focus
+        if (!this.open && !this.as<InputState>().isDisabled && !this.as<InputState>().isReadonly) { // open on focus
             this.open = true;
         }
 
-        this.as<InputManagement>().internalIsFocus = this.as<InputStateMixin>().active;
+        this.as<InputManagement>().internalIsFocus = !this.as<InputStateMixin>().isDisabled;
         if (this.as<InputManagement>().internalIsFocus) {
             this.$emit('focus', event);
         }
@@ -318,7 +321,7 @@ export class MDatepicker extends ModulVue {
 
     // override from InputManagement
     public onClick(event: MouseEvent): void {
-        this.as<InputManagement>().internalIsFocus = this.as<InputStateMixin>().active;
+        this.as<InputManagement>().internalIsFocus = !this.as<InputStateMixin>().isDisabled;
         let inputEl: HTMLElement | undefined = this.as<InputStateMixin>().getInput();
         if (this.as<InputManagement>().internalIsFocus && inputEl) {
             inputEl.focus();
