+++ packages/modul-components/src/components/textarea/textarea.html
@@ -24,7 +24,7 @@
         <textarea ref="input"
                   :id="id"
                   class="m-textarea__input"
-                  :disabled="isDisabled || isWaiting"
+                  :disabled="isDisabled"
                   :readonly="isReadonly"
                   :placeholder="placeholder"
                   :maxlength="maxLengthNumber"
