+++ packages/modul-components/src/components/textfield/textfield.html
@@ -28,7 +28,7 @@
                class="m-textfield__input"
                v-model="model"
                :autocomplete="autocomplete"
-               :disabled="isDisabled || isWaiting"
+               :disabled="isDisabled"
                :placeholder="placeholder"
                :readonly="isReadonly"
                :maxlength="maxLengthNumber"
